renamed aws modules
