+++ sentinelhub/aws/__init__.py
@@ -2,8 +2,8 @@
 The part of the package that implements AWS functionalities
 """
 
-from .aws import AwsProduct, AwsTile
-from .aws_safe import SafeProduct, SafeTile
 from .client import AwsDownloadClient
 from .constants import AwsConstants
-from .data_request import AwsProductRequest, AwsTileRequest, download_safe_format, get_safe_format
+from .data import AwsProduct, AwsTile
+from .data_safe import SafeProduct, SafeTile
+from .request import AwsProductRequest, AwsTileRequest, download_safe_format, get_safe_format
