+++ tests/aws/test_data_safe.py
